BookedChair oprettet

+++ src/main/java/stud/kea/dk/biografbackend/chair/model/ChairModel.java
@@ -16,7 +16,8 @@ public class ChairModel {
 
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long id;
+    private int id;
+
     private int chairNr;
     private int rowNr;
     private boolean isAvailable;
